+++ BOCCHI.Automator/IServiceCollectionExtensions.cs
@@ -3,12 +3,14 @@ using BOCCHI.Automator.Data.Goals;
 using BOCCHI.Automator.Services;
 using BOCCHI.Automator.Services.Goals;
 using BOCCHI.Automator.Services.Paths;
+using BOCCHI.Automator.Services.PotTreasure;
 using BOCCHI.Common;
 using BOCCHI.Common.Data.Goals;
 using BOCCHI.Common.Services;
 using BOCCHI.Common.Services.Paths;
 using Microsoft.Extensions.DependencyInjection;
 using Ocelot;
+using Ocelot.Lifecycle;
 using Ocelot.States;
 
 namespace BOCCHI.Automator;
@@ -25,6 +27,10 @@ public static class IServiceCollectionExtensions
         services.AddSingleton<IPotsTreasureMode, PotsTreasureService>();
         services.AddSingleton<Func<IPotsTreasureMode>>(sp => () => sp.GetRequiredService<IPotsTreasureMode>());
         services.AddSingleton<IllegalModeTreasureFillerService>();
+        services.AddSingleton<PotTreasureHintTracker>();
+        services.AddSingleton<IOnStart>(sp => sp.GetRequiredService<PotTreasureHintTracker>());
+        services.AddSingleton<IOnStop>(sp => sp.GetRequiredService<PotTreasureHintTracker>());
+        services.AddSingleton<MagicalElixirAssist>();
         services.AddSingleton<IDynamicRenderer, AutomatorRenderer>();
         services.AddSingleton<IDynamicRenderer, PotsTreasureRenderer>();